Avoid a slow invocation on each id links to speedup
Now we call `org-export--collect-tree-properties' on each id links,
and it's cost a lot of CPU. We now using a buffer-local variable to hold it
to avoid call it on each id links.

And also here is a little fix: before this PR in #453 we closed all
the `org-id-goto' buffers, and that cause the buffer we opened will
closed by us, and with this we'll only close the buffers that in
fundamental-mode.